+++ source/DSCResources/DSC_SqlProtocol/DSC_SqlProtocol.psm1
@@ -82,8 +82,11 @@ function Get-TargetResource
 
     $protocolNameProperties = Get-ProtocolNameProperties -ProtocolName $ProtocolName
 
+    # Getting the server protocol properties by using the computer name.
+    $computerName = Get-ComputerName
+
     Write-Verbose -Message (
-        $script:localizedData.GetCurrentState -f $protocolNameProperties.DisplayName, $InstanceName, $ServerName
+        $script:localizedData.GetCurrentState -f $protocolNameProperties.DisplayName, $InstanceName, $computerName
     )
 
     Import-SQLPSModule
@@ -93,7 +96,7 @@ function Get-TargetResource
         to a cluster instance or availability group listener.
     #>
     $getServerProtocolObjectParameters = @{
-        ServerName   = Get-ComputerName
+        ServerName   = $computerName
         Instance     = $InstanceName
         ProtocolName = $ProtocolName
     }
@@ -239,8 +242,11 @@ function Set-TargetResource
 
     if ($propertiesNotInDesiredState.Count -gt 0)
     {
+        # Getting the server protocol properties by using the computer name.
+        $computerName = Get-ComputerName
+
         Write-Verbose -Message (
-            $script:localizedData.SetDesiredState -f $protocolNameProperties.DisplayName, $InstanceName
+            $script:localizedData.SetDesiredState -f $protocolNameProperties.DisplayName, $InstanceName, $computerName
         )
 
         <#
@@ -248,7 +254,7 @@ function Set-TargetResource
             to a cluster instance or availability group listener.
         #>
         $getServerProtocolObjectParameters = @{
-            ServerName   = Get-ComputerName
+            ServerName   = $computerName
             Instance     = $InstanceName
             ProtocolName = $ProtocolName
         }
@@ -339,6 +345,10 @@ function Set-TargetResource
 
         if (-not $SuppressRestart -and $isRestartNeeded)
         {
+            <#
+                This is using the $ServerName to be able to restart a cluster
+                instance or availability group listener.
+            #>
             $restartSqlServiceParameters = @{
                 ServerName   = $ServerName
                 InstanceName = $InstanceName